And added the option to merge manifests.
Seems like all the other plugins always merge, which seems like a terrible policy to me.  But it's difficult to implement, and all these people spent the time to implement it and even made it the default, so I'll bow to their judgement and at least add it as an option.