Major refactoring of EclipseArgsBuilder to EclipseApp.
EclipseArgsBuilder was effectively modelling the input to an eclipsec instance.  When we started to build the ant integration, we needed to include the input build.xml into ArgsBuilder.  This made it clear we should rename EclispeArgsBuilder to EclipseApp, which would be responsible for running an eclipse utility application.

To abstract over the various mechanisms for running an EclipseApp (EquinoxLauncher, PdeBootstrapInstallation, etc), we introduced EclipseApp.Runner.